In-progress hooking the form provider into existing form
