+++ app/webpacker/components/CompetitionForm/FormSections/Admin.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import SubSection from './SubSection';
 import { InputBoolean } from '../Inputs/FormInputs';
 import { useStore } from '../../../lib/providers/StoreProvider';
+import SubSection from '../../wca/FormProvider/SubSection';
 
 export default function Admin() {
   const { isAdminView, isPersisted } = useStore();
